+++ src/Monitoring/TransactionMonitoring.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+namespace Kartenmacherei\RestFramework\Monitoring;
+
+interface TransactionMonitoring
+{
+    public function nameTransaction(string $transactionName);
+}