+++ src/Monitoring/VoidTransactionMonitoring.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+namespace Kartenmacherei\RestFramework\Monitoring;
+
+class VoidTransactionMonitoring implements TransactionMonitoring
+{
+    public function nameTransaction(string $transactionName)
+    {
+    }
+}